Refactor: `SimpleNav` page layout in own folder
