+++ src/modules/pages/components/RouteLayouts/SimpleNav/SimpleNav.css
