+++ src/modules/pages/components/RouteLayouts/SimpleNav/index.ts
@@ -1,0 +1,1 @@
+export { default } from './SimpleNav';
